% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatty.R
\name{chatty}
\alias{chatty}
\title{chatty}
\usage{
chatty(mydata, n)
}
\arguments{
\item{mydata}{my current datatable that has problem need to ask chatgpt}

\item{n}{how many rows i want to send to chatgpt}
}
\value{
a JSON object that chatgpt can receive
}
\description{
When i ask chatgpt about coding question, i need to give it a sample of  my data
}
\examples{
 chatty(iris, 1)
}
