% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.repo.R
\name{create.repo}
\alias{create.repo}
\title{create.repo}
\usage{
create.repo(proj_folder = here::here())
}
\arguments{
\item{proj_folder}{R project folder name, default here::here()}
}
\value{
a folder full with sub folders for R project developing use
}
\description{
Create a folder structure within your project path.including data, input, output, R etc.
}
\examples{
\dontrun{
create.repo("~/myproject")
}
}
