% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{create_subdirs}
\alias{create_subdirs}
\title{create_subdirs}
\usage{
create_subdirs(folder)
}
\arguments{
\item{folder}{path to the folder}
}
\value{
created some subfolder under the input folder
}
\description{
create_subdirs
}
\examples{
create_subdirs("C:/myproject")
}
