% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{create_subdirs_Boral}
\alias{create_subdirs_Boral}
\title{create_subdirs for Boreal project}
\usage{
create_subdirs_Boral(folder)
}
\arguments{
\item{folder}{path to the folder}
}
\value{
structured folder
}
\description{
create_subdirs for Boreal project
}
