% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifecycle.R
\name{lifecycle}
\alias{lifecycle}
\title{lifecycle}
\usage{
lifecycle(lifecycle)
}
\arguments{
\item{lifecycle}{a word choosing from experimental, maturing, stable, dormant, or retired}
}
\value{
the text you can copy and paste to readme.md
}
\description{
add the badge to your repo's readme doc.
}
\examples{
 lifecycle("experimental")
}
