% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepoBuild.R
\name{RepoBuild_Boreal}
\alias{RepoBuild_Boreal}
\title{RepoBuild Boreal}
\usage{
RepoBuild_Boreal(proj_folder)
}
\arguments{
\item{proj_folder}{your project folder}
}
\value{
Structured folder
}
\description{
Create a folder structure within your project path.Designed for Boreal
}
\examples{
 \dontrun{
RepoBuild_Boreal("C:/myproject1")
}
}
