% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{update_gitignore}
\alias{update_gitignore}
\title{update_gitignore}
\usage{
update_gitignore(folder)
}
\arguments{
\item{folder}{path to the project folder}
}
\value{
a new gitignore file with some default settings
}
\description{
update_gitignore
}
\examples{
update_gitignore("C:/myproject")
}
