% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.repo.R
\name{create.repo}
\alias{create.repo}
\title{create.repo}
\usage{
create.repo(proj_name = "abc9368-2022", proj_path = "C:/QuanGit")
}
\arguments{
\item{proj_name}{R project folder name}

\item{proj_path}{R project path, end without /}
}
\value{
a folder full with sub folders for R project developing use
}
\description{
create.repo
}
\examples{
create.repo(proj_name = "abc9368-2022", proj_path = "C:/Git")
}
