% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepoBuild.R
\name{RepoBuild}
\alias{RepoBuild}
\title{RepoBuild}
\usage{
RepoBuild(proj_folder)
}
\arguments{
\item{proj_folder}{Complete path to the R project folder with folder name}
}
\value{
a folder full with sub folders for R project developing use
}
\description{
Create a folder structure within your project path.including data, input, output, R etc.
}
\examples{
\dontrun{
RepoBuild("C:/myproject1")
}
}
